solidity: Added utils.js for a common place for testing helpers
